+++ esphome/components/substitutions/__init__.py
@@ -3,15 +3,15 @@ import logging
 from esphome import core
 from esphome.config_helpers import Extend, Remove, merge_config
 import esphome.config_validation as cv
-from esphome.const import CONF_JINJA, CONF_SUBSTITUTIONS, VALID_SUBSTITUTIONS_CHARACTERS
-from esphome.components.jinja import (
+from esphome.const import CONF_SUBSTITUTIONS, VALID_SUBSTITUTIONS_CHARACTERS
+from esphome.yaml_util import ESPHomeDataBase, make_data_base
+from .jinja import (
     Jinja,
     JinjaStr,
     has_jinja,
     TemplateError,
     TemplateRuntimeError,
 )
-from esphome.yaml_util import ESPHomeDataBase, make_data_base
 
 CODEOWNERS = ["@esphome/core"]
 _LOGGER = logging.getLogger(__name__)
@@ -167,11 +167,7 @@ def _substitute_item(substitutions, item, path, jinja, ignore_missing):
 
 
 def do_substitution_pass(config, command_line_substitutions, ignore_missing=False):
-    if (
-        CONF_SUBSTITUTIONS not in config
-        and not command_line_substitutions
-        and CONF_JINJA not in config
-    ):
+    if CONF_SUBSTITUTIONS not in config and not command_line_substitutions:
         return
 
     # Merge substitutions in config, overriding with substitutions coming from command line:
@@ -201,11 +197,5 @@ def do_substitution_pass(config, command_line_substitutions, ignore_missing=Fals
     config.move_to_end(CONF_SUBSTITUTIONS, False)
 
     # Create a Jinja environment that will consider substitutions in scope:
-    jinja = Jinja(config, substitutions)
-    jinja_config = None
-    if CONF_JINJA in config:
-        jinja_config = config[CONF_JINJA]
-        config[CONF_JINJA] = {}
+    jinja = Jinja(substitutions)
     _substitute_item(substitutions, config, [], jinja, ignore_missing)
-    if jinja_config is not None:
-        config[CONF_JINJA] = jinja_config
