+++ esphome/const.py
@@ -432,7 +432,6 @@ CONF_IP_ADDRESS = "ip_address"
 CONF_IRQ_PIN = "irq_pin"
 CONF_IS_RGBW = "is_rgbw"
 CONF_ITEMS = "items"
-CONF_JINJA = "jinja"
 CONF_JS_INCLUDE = "js_include"
 CONF_JS_URL = "js_url"
 CONF_JVC = "jvc"
