+++ tests/unit_tests/fixtures/substitutions/04-display_example.input.yaml
